refactor: UI 관련 객체를 view 패키지로 이동
